requirements.txt
discord.py==2.5.2
Flask==3.1.1
psutil==7.0.0
python-dotenv==1.1.0
yt_dlp==2025.5.22
